lab 4 final

+++ base/resources/record.py
@@ -2,6 +2,7 @@ import datetime
 import json
 import uuid
 from flask import Blueprint, jsonify, request
+from flask_jwt_extended import jwt_required, get_jwt_identity
 from ..schemas.Schemas import recordSchema
 from sqlalchemy.exc import IntegrityError
 from ..models import *
@@ -10,7 +11,9 @@ from ..db import db
 record_blueprint = Blueprint('record', __name__)
 
 @record_blueprint.post("/record")
+@jwt_required()
 def create_record():
+    user_id = get_jwt_identity()
     record_data = request.get_json()
     record_schema = recordSchema()
     try:
@@ -20,14 +23,13 @@ def create_record():
 
     validated_data["id"] = uuid.uuid4().hex
     validated_data["created_at"] = datetime.datetime.now()
-    user = userModel.query.get(record_data["user_id"])
     category = categoryModel.query.get(record_data["category_id"])
-    if(user and category):
-        validated_data["user_id"] = user.id
+    if category:
+        validated_data["user_id"] = user_id
         validated_data["category_id"] = category.id
         record = recordModel(**validated_data)
     else:
-        return "Incorrect record data", 400
+        return "Incorrect category id", 400
 
     try:
         db.session.add(record)
@@ -39,17 +41,16 @@ def create_record():
     return validated_data
 
 @record_blueprint.get("/record")
+@jwt_required()
 def get_records():
-    user_id = request.get_json()["user_id"]
-    category_id = request.get_json()["category_id"]
+    user_id = get_jwt_identity()
+    category_id = request.args.get("category_id")
     
     if user_id is None and category_id is None:
         return "Missing parameters", 400
     
     query = recordModel.query
 
-    if user_id:
-        query = query.filter_by(user_id=user_id)
     if category_id:
         query = query.filter_by(category_id=category_id)
 
@@ -58,11 +59,16 @@ def get_records():
     return record_schema.dump(records_list, many=True)
 
 @record_blueprint.delete("/record/<record_id>")
+@jwt_required()
 def record_delete(record_id):
+    user_id = get_jwt_identity()
     record = recordModel.query.get(record_id)
     if record:
-        db.session.delete(record)
-        db.session.commit()
-        return "", 204
+        if user_id == record.user_id:
+            db.session.delete(record)
+            db.session.commit()
+            return "", 204
+        else:
+            return "it`s not your record", 403
     else:
         return "Record not found", 404
